fixed the decimals not computing

+++ simplecalc.js
@@ -1,6 +1,6 @@
 function compute() {
-    var one = parseInt(document.getElementById("inputOne").value);
-    var two = parseInt(document.getElementById("inputTwo").value);
+    var one = parseFloat(document.getElementById("inputOne").value);
+    var two = parseFloat(document.getElementById("inputTwo").value);
     var sign = document.getElementById("operations").value;
     
     if (sign === "+") {
